# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "HMS API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/chat-consultations/create": {
      "post": {
        "summary": "Create a new chat session",
        "tags": [
          "ChatConsultation"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "patientId",
                  "doctorId"
                ],
                "properties": {
                  "patientId": {
                    "type": "string"
                  },
                  "doctorId": {
                    "type": "string"
                  },
                  "appointmentId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Chat session created"
          }
        }
      }
    },
    "/api/v1/chat-consultations/session/{sessionId}": {
      "get": {
        "summary": "Get details of a specific chat session",
        "tags": [
          "ChatConsultation"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "sessionId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session details fetched successfully"
          }
        }
      }
    },
    "/api/v1/chat-consultations/session/{sessionId}/messages": {
      "get": {
        "summary": "Get messages for a chat session",
        "tags": [
          "ChatConsultation"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "sessionId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Messages fetched successfully"
          }
        }
      }
    },
    "/api/v1/chat-consultations/send-message": {
      "post": {
        "summary": "Send a message in a chat session",
        "tags": [
          "ChatConsultation"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "sessionId",
                  "content"
                ],
                "properties": {
                  "sessionId": {
                    "type": "string"
                  },
                  "content": {
                    "type": "string"
                  },
                  "messageType": {
                    "type": "string",
                    "enum": [
                      "text",
                      "image",
                      "document"
                    ],
                    "default": "text"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Message sent"
          }
        }
      }
    },
    "/api/v1/chat-consultations/upload": {
      "post": {
        "summary": "Upload an image or document for chat",
        "tags": [
          "ChatConsultation"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "File uploaded successfully, returns URL"
          }
        }
      }
    },
    "/api/v1/chat-consultations/session/{sessionId}/end": {
      "put": {
        "summary": "End a chat session",
        "tags": [
          "ChatConsultation"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "sessionId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session ended"
          }
        }
      }
    },
...
    "/api/v1/finance/expenses": {
      "post": {
        "summary": "Record a new hospital expense",
        "tags": [
          "Finance"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "expenseName",
                  "category",
                  "amount"
                ],
                "properties": {
                  "expenseName": {
                    "type": "string"
                  },
                  "category": {
                    "type": "string",
                    "enum": [
                      "Utility Bill",
                      "Maintenance",
                      "Equipment Purchase",
                      "Medicine Purchase",
                      "Marketing",
                      "Miscellaneous",
                      "Other"
                    ]
                  },
                  "amount": {
                    "type": "number"
                  },
                  "dateIncurred": {
                    "type": "string",
                    "format": "date"
                  },
                  "description": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Expense recorded"
          }
        }
      },
      "get": {
        "summary": "Get hospital expenses",
        "tags": [
          "Finance"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "category",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "startDate",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "in": "query",
            "name": "endDate",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of expenses"
          }
        }
      }
    },
    "/api/v1/finance/payrolls": {
      "post": {
        "summary": "Create a salary/payroll record for a staff member",
        "tags": [
          "Finance"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "staff",
                  "salaryMonth",
                  "basicSalary"
                ],
                "properties": {
                  "staff": {
                    "type": "string"
                  },
                  "salaryMonth": {
                    "type": "string",
                    "example": "2023-10"
                  },
                  "basicSalary": {
                    "type": "number"
                  },
                  "bonus": {
                    "type": "number"
                  },
                  "deductions": {
                    "type": "number"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "Pending",
                      "Paid"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Payroll created"
          }
        }
      },
      "get": {
        "summary": "Get all payroll records",
        "tags": [
          "Finance"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "salaryMonth",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of payrolls"
          }
        }
      }
    },
    "/api/v1/finance/payrolls/{payrollId}/status": {
      "patch": {
        "summary": "Update payroll status (e.g. mark as Paid)",
        "tags": [
          "Finance"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "payrollId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "status"
                ],
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "Pending",
                      "Paid"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Payroll status updated"
          }
        }
      }
    },
    "/api/v1/finance/overview": {
      "get": {
        "summary": "Get Financial Overview (Profit/Loss statement)",
        "tags": [
          "Finance"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "startDate",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "in": "query",
            "name": "endDate",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Financial summary"
          }
        }
      }
    },
...
  },
...
}